+++ dataworkspace/dataworkspace/apps/applications/admin.py
@@ -19,7 +19,7 @@ from dataworkspace.apps.applications.models import (
     ApplicationTemplate,
     ToolTemplate,
     VisualisationTemplate,
-    Team
+
 )
 from dataworkspace.apps.datasets.models import (
     DataSet,
@@ -34,11 +34,6 @@ from dataworkspace.apps.applications.utils import (
 )
 from dataworkspace.apps.eventlog.models import EventLog
 from dataworkspace.apps.eventlog.utils import log_permission_change
-
-
-# @admin.register(Team)
-# class TeamAdmin(admin.ModelAdmin):
-#     pass
 
 
 @admin.register(ApplicationInstance)
